+++ core-ng/src/main/java/core/framework/api/util/Properties.java
@@ -23,8 +23,7 @@ public final class Properties {
             properties.load(reader);
             properties.forEach((key, value) -> {
                 String previous = this.properties.putIfAbsent((String) key, (String) value);
-                if (previous != null)
-                    throw Exceptions.error("property already exists, key={}, previous={}, current={}", key, previous, value);
+                if (previous != null) throw Exceptions.error("property already exists, key={}, previous={}, current={}", key, previous, value);
             });
         } catch (IOException e) {
             throw new UncheckedIOException(e);