+++ core-ng/src/main/java/core/framework/impl/log/ActionLog.java
@@ -1,10 +1,8 @@
 package core.framework.impl.log;
 
-import core.framework.api.log.Markers;
 import core.framework.api.util.Maps;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.slf4j.Marker;
 
 import java.time.Duration;
 import java.time.Instant;
@@ -18,7 +16,7 @@ import java.util.UUID;
  * @author neo
  */
 public final class ActionLog {
-    private static final int MAX_TRACE_HOLD_SIZE = 10000;
+    private static final int MAX_TRACE_HOLD_SIZE = 5000;
     public final String id = UUID.randomUUID().toString();
     final Instant startTime = Instant.now();
     final Map<String, String> context = Maps.newLinkedHashMap();
@@ -37,26 +35,23 @@ public final class ActionLog {
         if (event.level.value > result.value) {
             result = event.level;
             errorCode = event.errorCode(); // only update error type/message if level raised, so error type will be first WARN or first ERROR
-            errorMessage = event.message(200);  // limit 200 chars in action log
+            errorMessage = errorMessage(event);
         }
 
-        if (events.size() < MAX_TRACE_HOLD_SIZE) {
+        if (events.size() < MAX_TRACE_HOLD_SIZE || event.level.value >= LogLevel.WARN.value) {  // after reach max holding lines, only add warning/error events
             events.add(event);
         }
     }
 
-    void end() {
-        if (events.size() >= MAX_TRACE_HOLD_SIZE) {
-            Marker marker = Markers.errorCode("TRACE_LOG_TOO_LONG");
-            String message = "reached max holding size of trace log, please contact arch team";
-            if (result.value < LogLevel.WARN.value) {       // not hide existing warn/error if there is already one
-                result = LogLevel.WARN;
-                errorCode = marker.getName();
-                errorMessage = message;
-            }
-            LogEvent warning = new LogEvent(logger.getName(), marker, LogLevel.WARN, message, null, null);
-            events.add(warning);
+    private String errorMessage(LogEvent event) {
+        String message = event.message();
+        if (message.length() > 200) {
+            return message.substring(0, 200);    // limit 200 chars in action log
         }
+        return message;
+    }
+
+    void end() {
         elapsed = Duration.between(startTime, Instant.now()).toMillis();
     }