+++ core-ng/src/test/java/core/framework/impl/log/LogParamTest.java
@@ -1,5 +1,6 @@
 package core.framework.impl.log;
 
+import core.framework.api.util.Charsets;
 import core.framework.api.util.Maps;
 import core.framework.api.util.Strings;
 import org.junit.Test;
@@ -18,5 +19,12 @@ public class LogParamTest {
         values.put("key1", Strings.bytes("value1"));
         values.put("key2", Strings.bytes("value2"));
         assertEquals("{key1=value1, key2=value2}", String.valueOf(LogParam.of(values)));
+    }
+
+    @Test
+    public void truncateLongString() {
+        String message = "1234567890";
+        LogParam.StringParam param = new LogParam.StringParam(Strings.bytes(message), Charsets.UTF_8, 5);
+        assertEquals("1234...(truncated)", param.toString());
     }
 }